Deprecate SAMLAssertionClaimsCallback class and add DefaultOIDCClaimsCallbackHandler. Refactored SAMLAssertionClaimsCallback to remove duplicated logic.
